+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Auto-adapt number of processes for integration specs based on the number of CPUs
 - Introduce an integration spec runner that prints everything to stdout (better for development)
 - Introduce extra integration specs for various ActiveJob usage scenarios
+- Rename consumer method `#prepared` to `#prepare` to reflect better its use-case
 
 ## 2.0.0-beta2 (2022-06-07)
 - Abstract away notion of topics groups (until now it was just an array)